chore(s3): restructure xml parser
