+++ packages/storage/src/AwsClients/S3/utils/xmlParser.browser.ts
